requirements.txt
requests
astral
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

# Web API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
pydantic>=2.0.0
